Seller Approval Limit Is Updated

+++ src/SQL_DATABASE.sql
@@ -91,7 +91,7 @@ CREATE TABLE [sellerApprovalLimit]
 (
 	[id] int PRIMARY KEY IDENTITY(1000,1),
 	[employeeID] int FOREIGN KEY REFERENCES employee(id) ON DELETE CASCADE ON UPDATE CASCADE,
-	[maxApprovalLimit] decimal(10,2) NOT NULL
+	[maxApprovalLimit] int NOT NULL
 );
 
 INSERT INTO [sellerApprovalLimit] ([employeeID], [maxApprovalLimit]) VALUES
